+++ test/Main.hs
@@ -9,7 +9,11 @@ import Data.List (sort)
 import Test.Hspec
 
 import Streamly
+import Streamly.Prelude ((.:), nil)
 import qualified Streamly.Prelude as A
+
+singleton :: IsStream t => a -> t m a
+singleton a = a .: nil
 
 toListSerial :: SerialT IO a -> IO [a]
 toListSerial = A.toList . serially
@@ -20,8 +24,8 @@ toListInterleaved = A.toList . interleaving
 toListAsync :: AsyncT IO a -> IO [a]
 toListAsync = A.toList . asyncly
 
-toListParallel :: Ord a => ParallelT IO a -> IO [a]
-toListParallel = fmap sort . A.toList . parallely
+toListParallel :: ParallelT IO a -> IO [a]
+toListParallel = A.toList . parallely
 
 main :: IO ()
 main = hspec $ do
@@ -50,8 +54,9 @@ main = hspec $ do
         it "fmap on composed (<>)" $
             (toListSerial $ fmap (+1) (return 1 <> return 2))
                 `shouldReturn` ([2,3] :: [Int])
-        it "fmap on composed (<|>)" $
-            (toListSerial $ fmap (+1) (return 1 <|> return 2))
+
+        it "fmap on composed (<>)" $
+            ((toListParallel $ fmap (+1) (return 1 <> return 2)) >>= return .  sort)
                 `shouldReturn` ([2,3] :: [Int])
 
     ---------------------------------------------------------------------------
@@ -70,43 +75,43 @@ main = hspec $ do
                 `shouldReturn` ([(1,2),(1,3)] :: [(Int, Int)])
 
         it "Apply - parallel composed first argument" $
-            (toListSerial $ (,) <$> (return 1 <|> return 2) <*> (return 3))
+            (toListParallel ((,) <$> (return 1 <> return 2) <*> (return 3)) >>= return . sort)
                 `shouldReturn` ([(1,3),(2,3)] :: [(Int, Int)])
 
         it "Apply - parallel composed second argument" $
-            (toListSerial $ (,) <$> (return 1) <*> (return 2 <|> return 3))
+            (toListParallel ((,) <$> (return 1) <*> (return 2 <> return 3)) >>= return . sort)
                 `shouldReturn` ([(1,2),(1,3)] :: [(Int, Int)])
 
     ---------------------------------------------------------------------------
     -- Monoidal Compositions
     ---------------------------------------------------------------------------
 
-    describe "Serial Composition (<>)" $ compose (<>) id
-    describe "Serial Composition (mappend)" $ compose mappend id
-    describe "Interleaved Composition (<>)" $ compose (<=>) sort
-    describe "Left biased parallel Composition (<|)" $ compose (<|) sort
-    describe "Fair parallel Composition (<|>)" $ compose (<|>) sort
-    describe "Fair parallel Composition (mplus)" $ compose mplus sort
-
+    describe "Serial Composition" $ compose serially mempty id
+    describe "Interleaved Composition" $ compose interleaving mempty sort
+    describe "Left biased parallel Composition" $ compose asyncly mempty sort
+    describe "Fair parallel Composition" $ compose parallely mempty sort
+    describe "Semigroup Composition for ZipSerial" $ compose zipping mempty id
+    describe "Semigroup Composition for ZipAsync" $ compose zippingAsync mempty id
+    -- XXX need to check alternative compositions as well
     ---------------------------------------------------------------------------
     -- Monoidal Composition ordering checks
     ---------------------------------------------------------------------------
 
-    describe "Serial interleaved ordering check (<=>)" $ interleaveCheck (<=>)
-    describe "Parallel interleaved ordering check (<|>)" $ interleaveCheck (<|>)
-    describe "Left biased parallel time order check" $ parallelCheck (<|)
-    describe "Fair parallel time order check" $ parallelCheck (<|>)
+    describe "Serial interleaved ordering check" $ interleaveCheck interleaving
+    describe "Parallel interleaved ordering check" $ interleaveCheck parallely
+    describe "Left biased parallel time order check" $ parallelCheck asyncly
+    describe "Fair parallel time order check" $ parallelCheck parallely
 
     ---------------------------------------------------------------------------
     -- TBD Monoidal composition combinations
     ---------------------------------------------------------------------------
 
     -- TBD need more such combinations to be tested.
-    describe "<> and <>" $ composeAndComposeSimple (<>) (<>) (cycle [[1 .. 9]])
+    describe "<> and <>" $ composeAndComposeSimple serially serially (cycle [[1 .. 9]])
 
     describe "<> and <=>" $ composeAndComposeSimple
-      (<>)
-      (<=>)
+      serially
+      interleaving
       ([ [1 .. 9]
        , [1 .. 9]
        , [1, 3, 2, 4, 6, 5, 7, 9, 8]
@@ -114,8 +119,8 @@ main = hspec $ do
        ])
 
     describe "<=> and <=>" $ composeAndComposeSimple
-      (<=>)
-      (<=>)
+      interleaving
+      interleaving
       ([ [1, 4, 2, 7, 3, 5, 8, 6, 9]
        , [1, 7, 4, 8, 2, 9, 5, 3, 6]
        , [1, 4, 3, 7, 2, 6, 9, 5, 8]
@@ -123,8 +128,8 @@ main = hspec $ do
        ])
 
     describe "<=> and <>" $ composeAndComposeSimple
-      (<=>)
-      (<>)
+      interleaving
+      serially
       ([ [1, 4, 2, 7, 3, 5, 8, 6, 9]
        , [1, 7, 4, 8, 2, 9, 5, 3, 6]
        , [1, 4, 2, 7, 3, 5, 8, 6, 9]
@@ -132,6 +137,9 @@ main = hspec $ do
        ])
 
     describe "Nested parallel and serial compositions" $ do
+        let t = timed
+            p = adapt . parallely
+            s = adapt . serially
         {-
         -- This is not correct, the result can also be [4,4,8,0,8,0,2,2]
         -- because of parallelism of [8,0] and [8,0].
@@ -143,10 +151,9 @@ main = hspec $ do
             `shouldReturn` ([4,4,8,8,0,0,2,2])
         -}
         it "Nest <|>, <>, <|> (2)" $
-            let t = timed
-             in toListSerial (
-                    ((t 4 <|> t 8) <> (t 1 <|> t 2))
-                <|> ((t 4 <|> t 8) <> (t 1 <|> t 2)))
+            (A.toList . parallely) (
+                   s (p (t 4 <> t 8) <> p (t 1 <> t 2))
+                <> s (p (t 4 <> t 8) <> p (t 1 <> t 2)))
             `shouldReturn` ([4,4,8,8,1,1,2,2])
         -- FIXME: These two keep failing intermittently on Mac OS X
         -- Need to examine and fix the tests.
@@ -165,55 +172,82 @@ main = hspec $ do
             `shouldReturn` ([4,4,1,1,8,2,9,2])
         -}
         it "Nest <|>, <|>, <|>" $
-            let t = timed
-             in toListSerial (
-                    ((t 4 <|> t 8) <|> (t 0 <|> t 2))
-                <|> ((t 4 <|> t 8) <|> (t 0 <|> t 2)))
+            (A.toList . parallely) (
+                    ((t 4 <> t 8) <> (t 0 <> t 2))
+                <> ((t 4 <> t 8) <> (t 0 <> t 2)))
             `shouldReturn` ([0,0,2,2,4,4,8,8])
 
     ---------------------------------------------------------------------------
     -- Monoidal composition recursion loops
     ---------------------------------------------------------------------------
 
-    describe "Serial loops (<>)" $ loops (<>) id reverse
-    describe "Left biased parallel loops (<|)" $ loops (<|) sort sort
-    describe "Fair parallel loops (<|>)" $ loops (<|>) sort sort
+    describe "Serial loops (<>)" $ loops serially id reverse
+    describe "Left biased parallel loops (<|)" $ loops asyncly sort sort
+    describe "Fair parallel loops (<|>)" $ loops parallely sort sort
 
     ---------------------------------------------------------------------------
     -- Bind and monoidal composition combinations
     ---------------------------------------------------------------------------
 
-    forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-        describe "Bind and compose" $ bindAndComposeSimple toListSerial g
+    describe "Bind and compose1" $ bindAndComposeSimple serially serially
+    describe "Bind and compose2" $ bindAndComposeSimple serially interleaving
+    describe "Bind and compose3" $ bindAndComposeSimple serially asyncly
+    describe "Bind and compose4" $ bindAndComposeSimple serially parallely
 
-    forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-        describe "Bind and compose" $ bindAndComposeSimple toListInterleaved g
+    describe "Bind and compose1" $ bindAndComposeSimple interleaving serially
+    describe "Bind and compose2" $ bindAndComposeSimple interleaving interleaving
+    describe "Bind and compose3" $ bindAndComposeSimple interleaving asyncly
+    describe "Bind and compose4" $ bindAndComposeSimple interleaving parallely
 
-    forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-        describe "Bind and compose" $ bindAndComposeSimple toListAsync g
+    describe "Bind and compose1" $ bindAndComposeSimple asyncly serially
+    describe "Bind and compose2" $ bindAndComposeSimple asyncly interleaving
+    describe "Bind and compose3" $ bindAndComposeSimple asyncly asyncly
+    describe "Bind and compose4" $ bindAndComposeSimple asyncly parallely
 
-    forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-        describe "Bind and compose" $ bindAndComposeSimple toListParallel g
+    describe "Bind and compose1" $ bindAndComposeSimple parallely serially
+    describe "Bind and compose2" $ bindAndComposeSimple parallely interleaving
+    describe "Bind and compose3" $ bindAndComposeSimple parallely asyncly
+    describe "Bind and compose4" $ bindAndComposeSimple parallely parallely
 
-    let fldr f = foldr f empty
-        fldl f = foldl f empty
-     in do
-        forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-            forM_ [fldr, fldl] $ \k ->
-                describe "Bind and compose" $
-                    bindAndComposeHierarchy toListSerial (k g)
-        forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-            forM_ [fldr, fldl] $ \k ->
-                describe "Bind and compose" $
-                    bindAndComposeHierarchy toListInterleaved (k g)
-        forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-            forM_ [fldr, fldl] $ \k ->
-                describe "Bind and compose" $
-                    bindAndComposeHierarchy toListAsync (k g)
-        forM_ [(<>), (<=>), (<|), (<|>)] $ \g ->
-            forM_ [fldr, fldl] $ \k ->
-                describe "Bind and compose" $
-                    bindAndComposeHierarchy toListParallel (k g)
+    let fldr, fldl :: (IsStream t, Semigroup (t IO Int)) => [t IO Int] -> t IO Int
+        fldr = foldr (<>) nil
+        fldl = foldl (<>) nil
+
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy serially serially k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy serially interleaving k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy serially asyncly k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy serially parallely k
+
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy interleaving serially k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy interleaving interleaving k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy interleaving asyncly k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy interleaving parallely k
+
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy asyncly serially k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy asyncly interleaving k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy asyncly asyncly k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy asyncly parallely k
+
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy parallely serially k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy parallely interleaving k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy parallely asyncly k
+    forM_ [fldr, fldl] $ \k ->
+        describe "Bind and compose" $ bindAndComposeHierarchy parallely parallely k
 
     -- Nest two lists using different styles of product compositions
     it "Nests two streams using monadic serial composition" nestTwoSerial
@@ -229,8 +263,7 @@ main = hspec $ do
     it "Nests two streams using Num serial composition" nestTwoSerialNum
     it "Nests two streams using Num interleaved composition" nestTwoInterleavedNum
     it "Nests two streams using Num async composition" nestTwoAsyncNum
-    -- This test fails intermittently, need to investigate
-    -- it "Nests two streams using Num parallel composition" nestTwoParallelNum
+    it "Nests two streams using Num parallel composition" nestTwoParallelNum
 
     ---------------------------------------------------------------------------
     -- TBD Bind and Bind combinations
@@ -291,182 +324,183 @@ nestTwoAsync :: Expectation
 nestTwoAsync =
     let s1 = foldMapWith (<>) return [1..4]
         s2 = foldMapWith (<>) return [5..8]
-    in toListAsync (do
+    in (toListAsync (do
         x <- s1
         y <- s2
         return (x + y)
-        ) `shouldReturn` ([6,7,8,9,7,8,9,10,8,9,10,11,9,10,11,12] :: [Int])
+        ) >>= return . sort)
+    `shouldReturn` sort ([6,7,8,9,7,8,9,10,8,9,10,11,9,10,11,12] :: [Int])
 
 nestTwoAsyncApp :: Expectation
 nestTwoAsyncApp =
     let s1 = foldMapWith (<>) return [1..4]
         s2 = foldMapWith (<>) return [5..8]
-    in toListAsync ((+) <$> s1 <*> s2)
-        `shouldReturn` ([6,7,8,9,7,8,9,10,8,9,10,11,9,10,11,12] :: [Int])
+    in (toListAsync ((+) <$> s1 <*> s2) >>= return . sort)
+        `shouldReturn` sort ([6,7,8,9,7,8,9,10,8,9,10,11,9,10,11,12] :: [Int])
 
 nestTwoAsyncNum :: Expectation
 nestTwoAsyncNum =
     let s1 = foldMapWith (<>) return [1..4]
         s2 = foldMapWith (<>) return [5..8]
-    in toListAsync (s1 + s2)
-        `shouldReturn` ([6,7,8,9,7,8,9,10,8,9,10,11,9,10,11,12] :: [Int])
+    in (toListAsync (s1 + s2) >>= return . sort)
+        `shouldReturn` sort ([6,7,8,9,7,8,9,10,8,9,10,11,9,10,11,12] :: [Int])
 
 nestTwoParallel :: Expectation
 nestTwoParallel =
     let s1 = foldMapWith (<>) return [1..4]
         s2 = foldMapWith (<>) return [5..8]
-    in toListParallel (do
+    in (toListParallel (do
         x <- s1
         y <- s2
         return (x + y)
-        ) `shouldReturn` ([6,7,7,8,8,8,9,9,9,9,10,10,10,11,11,12] :: [Int])
+        ) >>= return . sort)
+    `shouldReturn` sort ([6,7,7,8,8,8,9,9,9,9,10,10,10,11,11,12] :: [Int])
 
 nestTwoParallelApp :: Expectation
 nestTwoParallelApp =
     let s1 = foldMapWith (<>) return [1..4]
         s2 = foldMapWith (<>) return [5..8]
-    in toListParallel ((+) <$> s1 <*> s2)
-        `shouldReturn` ([6,7,7,8,8,8,9,9,9,9,10,10,10,11,11,12] :: [Int])
+    in (toListParallel ((+) <$> s1 <*> s2) >>= return . sort)
+        `shouldReturn` sort ([6,7,7,8,8,8,9,9,9,9,10,10,10,11,11,12] :: [Int])
 
-{-
 nestTwoParallelNum :: Expectation
 nestTwoParallelNum =
     let s1 = foldMapWith (<>) return [1..4]
         s2 = foldMapWith (<>) return [5..8]
-    in toListParallel (s1 + s2)
-        `shouldReturn` ([6,7,7,8,8,8,9,9,9,9,10,10,10,11,11,12] :: [Int])
--}
+    in (toListParallel (s1 + s2) >>= return . sort)
+        `shouldReturn` sort ([6,7,7,8,8,8,9,9,9,9,10,10,10,11,11,12] :: [Int])
 
-timed :: Int -> SerialT IO Int
+timed :: MonadIO (t IO) => Int -> t IO Int
 timed x = liftIO (threadDelay (x * 100000)) >> return x
 
-interleaveCheck
-    :: (SerialT IO Int -> SerialT IO Int -> SerialT IO Int)
-    -> Spec
-interleaveCheck f =
+interleaveCheck :: (IsStream t, Semigroup (t IO Int))
+    => (t IO Int -> t IO Int) -> Spec
+interleaveCheck t =
     it "Interleave four" $
-        toListSerial ((return 0 <> return 1) `f` (return 100 <> return 101))
+        (A.toList . t) ((singleton 0 <> singleton 1) <> (singleton 100 <> singleton 101))
             `shouldReturn` ([0, 100, 1, 101])
 
-parallelCheck :: (SerialT IO Int -> SerialT IO Int -> SerialT IO Int) -> Spec
-parallelCheck f = do
+parallelCheck :: (IsStream t, Semigroup (t IO Int), MonadIO (t IO))
+    => (t IO Int -> t IO Int) -> Spec
+parallelCheck t = do
     it "Parallel ordering left associated" $
-        toListSerial (((event 4 `f` event 3) `f` event 2) `f` event 1)
+        (A.toList . t) (((event 4 <> event 3) <> event 2) <> event 1)
             `shouldReturn` ([1..4])
 
     it "Parallel ordering right associated" $
-        toListSerial (event 4 `f` (event 3 `f` (event 2 `f` event 1)))
+        (A.toList . t) (event 4 <> (event 3 <> (event 2 <> event 1)))
             `shouldReturn` ([1..4])
 
     where event n = (liftIO $ threadDelay (n * 100000)) >> (return n)
 
-compose
-    :: (SerialT IO Int -> SerialT IO Int -> SerialT IO Int)
-    -> ([Int] -> [Int])
-    -> Spec
-compose f srt = do
+compose :: (IsStream t, Semigroup (t IO Int))
+    => (t IO Int -> t IO Int) -> t IO Int -> ([Int] -> [Int]) -> Spec
+compose t z srt = do
+    -- XXX these should get covered by the property tests
     it "Compose mempty, mempty" $
-        (tl (mempty `f` mempty)) `shouldReturn` []
-    it "Compose empty, empty" $
-        (tl (empty `f` empty)) `shouldReturn` []
+        (tl (z <> z)) `shouldReturn` ([] :: [Int])
     it "Compose empty at the beginning" $
-        (tl $ (empty `f` return 1)) `shouldReturn` [1]
+        (tl $ (z <> singleton 1)) `shouldReturn` [1]
     it "Compose empty at the end" $
-        (tl $ (return 1 `f` empty)) `shouldReturn` [1]
+        (tl $ (singleton 1 <> z)) `shouldReturn` [1]
     it "Compose two" $
-        (tl (return 0 `f` return 1) >>= return . srt)
+        (tl (singleton 0 <> singleton 1) >>= return . srt)
             `shouldReturn` [0, 1]
+    it "Compose many" $
+        ((tl $ forEachWith (<>) [1..100] singleton) >>= return . srt)
+            `shouldReturn` [1..100]
+
+    -- These are not covered by the property tests
     it "Compose three - empty in the middle" $
-        ((tl $ (return 0 `f` empty `f` return 1)) >>= return . srt)
+        ((tl $ (singleton 0 <> z <> singleton 1)) >>= return . srt)
             `shouldReturn` [0, 1]
     it "Compose left associated" $
-        ((tl $ (((return 0 `f` return 1) `f` return 2) `f` return 3))
+        ((tl $ (((singleton 0 <> singleton 1) <> singleton 2) <> singleton 3))
             >>= return . srt) `shouldReturn` [0, 1, 2, 3]
     it "Compose right associated" $
-        ((tl $ (return 0 `f` (return 1 `f` (return 2 `f` return 3))))
+        ((tl $ (singleton 0 <> (singleton 1 <> (singleton 2 <> singleton 3))))
             >>= return . srt) `shouldReturn` [0, 1, 2, 3]
-    it "Compose many" $
-        ((tl $ forEachWith f [1..100] return) >>= return . srt)
-            `shouldReturn` [1..100]
     it "Compose hierarchical (multiple levels)" $
-        ((tl $ (((return 0 `f` return 1) `f` (return 2 `f` return 3))
-                `f` ((return 4 `f` return 5) `f` (return 6 `f` return 7)))
+        ((tl $ (((singleton 0 <> singleton 1) <> (singleton 2 <> singleton 3))
+                <> ((singleton 4 <> singleton 5) <> (singleton 6 <> singleton 7)))
             ) >>= return . srt) `shouldReturn` [0..7]
-    where tl = toListSerial
+    where tl = A.toList . t
 
 composeAndComposeSimple
-    :: (SerialT IO Int -> SerialT IO Int -> SerialT IO Int)
-    -> (SerialT IO Int -> SerialT IO Int -> SerialT IO Int)
-    -> [[Int]]
-    -> Spec
-composeAndComposeSimple f g answer = do
+    :: ( IsStream t1, Semigroup (t1 IO Int)
+       , IsStream t2, Semigroup (t2 IO Int), Monoid (t2 IO Int), Monad (t2 IO))
+    => (t1 IO Int -> t1 IO Int)
+    -> (t2 IO Int -> t2 IO Int)
+    -> [[Int]] -> Spec
+composeAndComposeSimple t1 t2 answer = do
+    let rfold = adapt . t2 . foldMapWith (<>) return
     it "Compose right associated outer expr, right folded inner" $
-        let fold = foldMapWith g return
-         in (toListSerial (fold [1,2,3] `f` (fold [4,5,6] `f` fold [7,8,9])))
+         ((A.toList. t1) (rfold [1,2,3] <> (rfold [4,5,6] <> rfold [7,8,9])))
             `shouldReturn` (answer !! 0)
 
     it "Compose left associated outer expr, right folded inner" $
-        let fold = foldMapWith g return
-         in (toListSerial ((fold [1,2,3] `f` fold [4,5,6]) `f` fold [7,8,9]))
+         ((A.toList . t1) ((rfold [1,2,3] <> rfold [4,5,6]) <> rfold [7,8,9]))
             `shouldReturn` (answer !! 1)
 
+    let lfold xs = adapt $ t2 $ foldl (<>) mempty $ map return xs
     it "Compose right associated outer expr, left folded inner" $
-        let fold xs = foldl g empty $ map return xs
-         in (toListSerial (fold [1,2,3] `f` (fold [4,5,6] `f` fold [7,8,9])))
+         ((A.toList . t1) (lfold [1,2,3] <> (lfold [4,5,6] <> lfold [7,8,9])))
             `shouldReturn` (answer !! 2)
 
     it "Compose left associated outer expr, left folded inner" $
-        let fold xs = foldl g empty $ map return xs
-         in (toListSerial ((fold [1,2,3] `f` fold [4,5,6]) `f` fold [7,8,9]))
+         ((A.toList . t1) ((lfold [1,2,3] <> lfold [4,5,6]) <> lfold [7,8,9]))
             `shouldReturn` (answer !! 3)
 
-
 loops
-    :: (SerialT IO Int -> SerialT IO Int -> SerialT IO Int)
+    :: (IsStream t, Semigroup (t IO Int), MonadIO (t IO))
+    => (t IO Int -> t IO Int)
     -> ([Int] -> [Int])
     -> ([Int] -> [Int])
     -> Spec
-loops f tsrt hsrt = do
-    it "Tail recursive loop" $ (toListSerial (loopTail 0) >>= return . tsrt)
+loops t tsrt hsrt = do
+    it "Tail recursive loop" $ (A.toList (loopTail 0) >>= return . tsrt)
             `shouldReturn` [0..3]
 
-    it "Head recursive loop" $ (toListSerial (loopHead 0) >>= return . hsrt)
+    it "Head recursive loop" $ (A.toList (loopHead 0) >>= return . hsrt)
             `shouldReturn` [0..3]
 
     where
         loopHead x = do
             -- this print line is important for the test (causes a bind)
             liftIO $ putStrLn "LoopHead..."
-            (if x < 3 then loopHead (x + 1) else empty) `f` return x
+            t $ (if x < 3 then loopHead (x + 1) else nil) <> return x
 
         loopTail x = do
             -- this print line is important for the test (causes a bind)
             liftIO $ putStrLn "LoopTail..."
-            return x `f` (if x < 3 then loopTail (x + 1) else empty)
+            t $ return x <> (if x < 3 then loopTail (x + 1) else nil)
 
 bindAndComposeSimple
-    :: (IsStream t, Alternative (t IO), Monad (t IO))
-    => (forall a. Ord a => t IO a -> IO [a])
-    -> (t IO Int -> t IO Int -> t IO Int)
+    :: ( IsStream t1, IsStream t2, Semigroup (t2 IO Int), Monad (t2 IO))
+    => (t1 IO Int -> t1 IO Int)
+    -> (t2 IO Int -> t2 IO Int)
     -> Spec
-bindAndComposeSimple tl g = do
+bindAndComposeSimple t1 t2 = do
+    -- XXX need a bind in the body of forEachWith instead of a simple return
     it "Compose many (right fold) with bind" $
-        (tl (forEachWith g [1..10 :: Int] $ \x -> return x `f` (return .  id))
+        ((A.toList . t1) (adapt . t2 $ forEachWith (<>) [1..10 :: Int] return)
             >>= return . sort) `shouldReturn` [1..10]
 
     it "Compose many (left fold) with bind" $
-        let forL xs k = foldl g empty $ map k xs
-         in (tl (forL [1..10 :: Int] $ \x -> return x `f` (return . id))
+        let forL xs k = foldl (<>) nil $ map k xs
+         in ((A.toList . t1) (adapt . t2 $ forL [1..10 :: Int] return)
                 >>= return . sort) `shouldReturn` [1..10]
-    where f = (>>=)
 
 bindAndComposeHierarchy
-    :: Monad (s IO) => (forall a. Ord a => s IO a -> IO [a])
-    -> ([s IO Int] -> s IO Int)
+    :: ( IsStream t1, Monad (t1 IO)
+       , IsStream t2, Monad (t2 IO))
+    => (t1 IO Int -> t1 IO Int)
+    -> (t2 IO Int -> t2 IO Int)
+    -> ([t2 IO Int] -> t2 IO Int)
     -> Spec
-bindAndComposeHierarchy tl g = do
+bindAndComposeHierarchy t1 t2 g = do
     it "Bind and compose nested" $
-        (tl bindComposeNested >>= return . sort)
+        ((A.toList . t1) bindComposeNested >>= return . sort)
             `shouldReturn` (sort (
                    [12, 18]
                 ++ replicate 3 13
@@ -489,12 +523,11 @@ bindAndComposeHierarchy tl g = do
 --         in m
          in b
 
-    tripleCompose a b c = g [a, b, c]
+    tripleCompose a b c = adapt . t2 $ g [a, b, c]
     tripleBind mx my mz =
-        mx `f` \x -> my
-           `f` \y -> mz
-           `f` \z -> return (x + y + z)
-    f = (>>=)
+        mx >>= \x -> my
+           >>= \y -> mz
+           >>= \z -> return (x + y + z)
 
 mixedOps :: Spec
 mixedOps = do
@@ -512,14 +545,14 @@ mixedOps = do
         x <- return 1
         y <- return 2
         z <- do
-                x1 <- return 1 <|> return 2
+                x1 <- adapt . parallely $ return 1 <> return 2
                 liftIO $ return ()
                 liftIO $ putStr ""
-                y1 <- return 1 <| return 2
+                y1 <- adapt . asyncly $ return 1 <> return 2
                 z1 <- do
                     x11 <- return 1 <> return 2
-                    y11 <- return 1 <| return 2
-                    z11 <- return 1 <=> return 2
+                    y11 <- adapt . asyncly $ return 1 <> return 2
+                    z11 <- adapt . interleaving $ return 1 <> return 2
                     liftIO $ return ()
                     liftIO $ putStr ""
                     return (x11 + y11 + z11)